CTCM/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy<2.0.0  # PyTorch compatibility
scipy>=1.7.0

# Model architecture
einops>=0.6.0
einops-exts>=0.0.4

# Configuration and logging
omegaconf>=2.3.0
hydra-core>=1.3.0
wandb>=0.15.0
tensorboard>=2.11.0

# Data and augmentation
Pillow>=9.0.0
torchdata>=0.6.0

# Training utilities
tqdm>=4.65.0
accelerate>=0.20.0

# Evaluation metrics
pytorch-fid>=0.3.0
lpips>=0.1.4
torchmetrics>=0.11.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Optional: For distributed training
# horovod>=0.27.0
# deepspeed>=0.9.0

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# isort>=5.10.0
# flake8>=4.0.0
# mypy>=0.950
# pre-commit>=2.17.0